Fix custom version prompt (#769)
The validate function fails when any string is returned. The return of semver.valid will be the same version string (as long as it's valid). So, a valid version will be an "error" according to the validate function. Switching to check for a non-null value coerces this to a boolean that validate is looking for.